feat: Refactor helpMenu Modal to seperate package
