+++ src/internal/handle_modal.go
@@ -6,7 +6,6 @@ import (
 	"path/filepath"
 	"strings"
 
-	"github.com/yorukot/superfile/src/internal/common"
 	"github.com/yorukot/superfile/src/internal/ui/filepanel"
 	"github.com/yorukot/superfile/src/internal/utils"
 )
@@ -104,156 +103,6 @@ func (m *model) cancelSearch() {
 func (m *model) confirmSearch() {
 	panel := m.getFocusedFilePanel()
 	panel.SearchBar.Blur()
-}
-
-// Help menu panel list up
-func (m *model) helpMenuListUp() {
-	if m.helpMenu.cursor > 1 {
-		m.helpMenu.cursor--
-		if m.helpMenu.cursor < m.helpMenu.renderIndex {
-			m.helpMenu.renderIndex = m.helpMenu.cursor
-		}
-		if m.helpMenu.filteredData[m.helpMenu.cursor].subTitle != "" {
-			m.helpMenu.cursor--
-		}
-	} else {
-		// Set the cursor to the last item in the list.
-		// We use max(..., 0) as a safeguard to prevent a negative cursor index
-		// in case the filtered list is empty.
-		m.helpMenu.cursor = max(len(m.helpMenu.filteredData)-1, 0)
-
-		// Adjust the render index to show the bottom of the list.
-		// Similarly, we use max(..., 0) to ensure the renderIndex doesn't become negative,
-		// which can happen if the number of items is less than the view height.
-		// This prevents a potential out-of-bounds panic during rendering.
-		m.helpMenu.renderIndex = max(len(m.helpMenu.filteredData)-(m.helpMenu.height-common.InnerPadding), 0)
-	}
-}
-
-// Help menu panel list down
-func (m *model) helpMenuListDown() {
-	if len(m.helpMenu.filteredData) == 0 {
-		return
-	}
-
-	if m.helpMenu.cursor < len(m.helpMenu.filteredData)-1 {
-		// Compute the next selectable row (skip subtitles).
-		next := m.helpMenu.cursor + 1
-		for next < len(m.helpMenu.filteredData) && m.helpMenu.filteredData[next].subTitle != "" {
-			next++
-		}
-		if next >= len(m.helpMenu.filteredData) {
-			// Wrap if no more selectable rows.
-			m.helpMenu.cursor = 1
-			m.helpMenu.renderIndex = 0
-			return
-		}
-		m.helpMenu.cursor = next
-
-		// Scroll down if cursor moved past the viewport.
-		if m.helpMenu.cursor > m.helpMenu.renderIndex+m.helpMenu.height-5 {
-			m.helpMenu.renderIndex++
-		}
-		// Clamp renderIndex to bottom.
-		bottom := len(m.helpMenu.filteredData) - (m.helpMenu.height - common.InnerPadding)
-		if bottom < 0 {
-			bottom = 0
-		}
-		if m.helpMenu.renderIndex > bottom {
-			m.helpMenu.renderIndex = bottom
-		}
-	} else {
-		m.helpMenu.cursor = 1
-		m.helpMenu.renderIndex = 0
-	}
-}
-
-func removeOrphanSections(items []helpMenuModalData) []helpMenuModalData {
-	var result []helpMenuModalData
-	// Since we can't know beforehand which section are we actually filtering
-	// we may end up in a scenario where there are two sections (General, Panel navigation)
-	// with no hotkeys between them, so we need to remove the section which its hotkeys was
-	// completely filtered out (Orphan sections)
-	for i := range items {
-		if items[i].subTitle != "" {
-			// Look ahead: is the next item a real hotkey?
-			if i+1 < len(items) && items[i+1].subTitle == "" {
-				result = append(result, items[i])
-			}
-			// Else: skip this subtitle because no children
-		} else {
-			result = append(result, items[i])
-		}
-	}
-	return result
-}
-
-func (m *model) filterHelpMenu(query string) {
-	filtered := fuzzySearch(query, m.helpMenu.data)
-	filtered = removeOrphanSections(filtered)
-
-	m.helpMenu.filteredData = filtered
-	if len(filtered) == 0 {
-		m.helpMenu.cursor = 0
-	} else {
-		m.helpMenu.cursor = 1
-	}
-	m.helpMenu.renderIndex = 0
-}
-
-// Fuzzy search function for a list of helpMenuModalData.
-// inspired from: sidebar/directory_utils.go
-func fuzzySearch(query string, data []helpMenuModalData) []helpMenuModalData {
-	if len(data) == 0 {
-		return []helpMenuModalData{}
-	}
-
-	// Optimization - This haystack can be kept precomputed based on description
-	// instead of re computing it in each call
-	haystack := []string{}
-	idxMap := []int{}
-	for i, item := range data {
-		if item.subTitle != "" {
-			continue
-		}
-		searchText := strings.Join(item.hotkey, " ") + " " + item.description
-		haystack = append(haystack, searchText)
-		idxMap = append(idxMap, i)
-	}
-
-	matchedIdx := map[int]struct{}{}
-	for _, match := range utils.FzfSearch(query, haystack) {
-		matchedIdx[idxMap[match.HayIndex]] = struct{}{}
-	}
-
-	results := []helpMenuModalData{}
-	for i, d := range data {
-		_, isMatch := matchedIdx[i]
-		if d.subTitle != "" || isMatch {
-			results = append(results, d)
-		}
-	}
-
-	return results
-}
-
-// Toggle help menu
-func (m *model) openHelpMenu() {
-	if m.helpMenu.open {
-		m.helpMenu.searchBar.Reset()
-		m.helpMenu.open = false
-		return
-	}
-
-	// Reset filteredData to the full data whenever the helpMenu is opened
-	m.helpMenu.filteredData = m.helpMenu.data
-	m.helpMenu.open = true
-}
-
-// Quit help menu
-func (m *model) quitHelpMenu() {
-	m.helpMenu.searchBar.Reset()
-	m.helpMenu.open = false
 }
 
 func (m *model) getFocusedFilePanel() *filepanel.Model {
