+++ src/internal/model_render.go
@@ -1,12 +1,9 @@
 package internal
 
 import (
-	"fmt"
 	"path/filepath"
 	"strconv"
 
-	"github.com/yorukot/superfile/src/internal/ui"
-	"github.com/yorukot/superfile/src/internal/ui/rendering"
 	filepreview "github.com/yorukot/superfile/src/pkg/file_preview"
 
 	"github.com/yorukot/superfile/src/internal/common"
@@ -119,7 +116,7 @@ func (m *model) introduceModalRender() string {
 			"      https://github.com/yorukot/superfile\n"+
 			"      Of course, you can always open a new issue to share your idea \n"+
 			"      or report a bug!")
-	return common.FirstUseModal(m.helpMenu.height, m.helpMenu.width).
+	return common.FirstUseModal(m.helpMenu.GetHeight(), m.helpMenu.GetWidth()).
 		Render(title + "\n\n" + vimUserWarn + "\n\n" + subOne + "\n\n" +
 			subTwo + "\n\n" + subThree + "\n\n" + subFour + "\n\n")
 }
@@ -130,87 +127,4 @@ func (m *model) promptModalRender() string {
 
 func (m *model) zoxideModalRender() string {
 	return m.zoxideModal.Render()
-}
-
-func (m *model) helpMenuRender() string {
-	r := ui.HelpMenuRenderer(m.helpMenu.height, m.helpMenu.width)
-	r.AddLines(" " + m.helpMenu.searchBar.View())
-	r.AddLines("") // one-line separation between searchbar and content
-
-	// TODO : This computation should not happen at render time. Move this to update
-	// TODO : Move these computations to a utility function
-	maxKeyLength := 0
-	for _, data := range m.helpMenu.filteredData {
-		totalKeyLen := 0
-		for _, key := range data.hotkey {
-			totalKeyLen += len(key)
-		}
-
-		separatorLen := max(0, (len(data.hotkey)-1)) * common.FooterGroupCols
-		if data.subTitle == "" && totalKeyLen+separatorLen > maxKeyLength {
-			maxKeyLength = totalKeyLen + separatorLen
-		}
-	}
-
-	valueLength := m.helpMenu.width - maxKeyLength - common.BorderPadding
-	if valueLength < m.helpMenu.width/common.CenterDivisor {
-		valueLength = m.helpMenu.width/common.CenterDivisor - common.BorderPadding
-	}
-
-	totalTitleCount := 0
-	cursorBeenTitleCount := 0
-
-	for i, data := range m.helpMenu.filteredData {
-		if data.subTitle != "" {
-			if i < m.helpMenu.cursor {
-				cursorBeenTitleCount++
-			}
-			totalTitleCount++
-		}
-	}
-
-	renderHotkeyLength := m.getRenderHotkeyLengthHelpMenuModal()
-	m.getHelpMenuContent(r, renderHotkeyLength, valueLength)
-
-	current := m.helpMenu.cursor + 1 - cursorBeenTitleCount
-	if len(m.helpMenu.filteredData) == 0 {
-		current = 0
-	}
-	r.SetBorderInfoItems(fmt.Sprintf("%s/%s",
-		strconv.Itoa(current),
-		strconv.Itoa(len(m.helpMenu.filteredData)-totalTitleCount)))
-	return r.Render()
-}
-
-func (m *model) getRenderHotkeyLengthHelpMenuModal() int {
-	renderHotkeyLength := 0
-	for i := m.helpMenu.renderIndex; i < m.helpMenu.renderIndex+(m.helpMenu.height-common.InnerPadding) && i < len(m.helpMenu.filteredData); i++ {
-		if m.helpMenu.filteredData[i].subTitle != "" {
-			continue
-		}
-
-		hotkey := common.GetHelpMenuHotkeyString(m.helpMenu.filteredData[i].hotkey)
-
-		renderHotkeyLength = max(renderHotkeyLength, len(common.HelpMenuHotkeyStyle.Render(hotkey)))
-	}
-	return renderHotkeyLength + 1
-}
-
-func (m *model) getHelpMenuContent(r *rendering.Renderer, renderHotkeyLength int, valueLength int) {
-	for i := m.helpMenu.renderIndex; i < m.helpMenu.renderIndex+(m.helpMenu.height-common.InnerPadding) && i < len(m.helpMenu.filteredData); i++ {
-		if m.helpMenu.filteredData[i].subTitle != "" {
-			r.AddLines(common.HelpMenuTitleStyle.Render(" " + m.helpMenu.filteredData[i].subTitle))
-			continue
-		}
-
-		hotkey := common.GetHelpMenuHotkeyString(m.helpMenu.filteredData[i].hotkey)
-		description := common.TruncateText(m.helpMenu.filteredData[i].description, valueLength, "...")
-
-		cursor := "  "
-		if m.helpMenu.cursor == i {
-			cursor = common.FilePanelCursorStyle.Render(icon.Cursor + " ")
-		}
-		r.AddLines(cursor + common.ModalStyle.Render(fmt.Sprintf("%*s%s", renderHotkeyLength,
-			common.HelpMenuHotkeyStyle.Render(hotkey+" "), common.ModalStyle.Render(description))))
-	}
 }
